Improve error handling
The Magento command at https://github.com/sonassi/magestack-scripts/blob/master/cli/magento-2/cron.sh#L43 can fail to run owing to a lack of memory, yet as this returns an exit code of 0, the wrapper does not send an email alert.

Additionally, if something goes wrong with the cron the exit code of the wrapper itself is 0 (indicating success), when it should be 1 (indicating an error). 

This pull request prevents failure-appearing-as-success from happening, and is flexible enough for a user to specify any wording in the log file that would indicate a cron had not run successfully.

Additionally, on discovering an error the wrapper itself will exit with an error code, ensuring any chained or piped commands (e.g. a monitoring service) are aware there has been an error.